Выносит функцию cut_string и mock-данные в отдельные папки
