+++ mock/is_auth.php
@@ -1,0 +1,2 @@
+<?php
+$is_auth = rand(0, 1);
