+++ mock/user_name.php
@@ -1,0 +1,7 @@
+<?php
+/**
+ * Захардкоженное имя пользователя.
+ * @global string $GLOBALS ['user_name']
+ * @name $user_name
+ */
+$user_name = 'Данил';
